Account for empty responses from rest API
Sometimes, there is no response during a successful operation from the
JIRA rest API, and this should handle that in the short term.  A better
solution might be to check the status code of the response, and only
check for errors if it is a 4XX or 5XX status code.  This is an attempt
to resolve issue #20