New sprint issue end point?!
I was trying to use the `addIssueToSprint` function and kept getting 404's back from the JIRA API.  After some digging, it looks like it's now located in the agile api with a few minor changes.  I was able to successfully add an issue to a sprint this these updates.

Heres the example from the updated docs:
```
curl --request POST \
  --user email@example.com:<api_token> \
  --header 'Accept: application/json' \
  --header 'Content-Type: application/json' \
  --data '
    {
      "issues": [
        "PR-1",
        "10001",
        "PR-3"
      ]
    }' \
  --url 'https://your-domain.atlassian.net/rest/agile/1.0/sprint/{sprintId}/issue'
```
https://developer.atlassian.com/cloud/jira/software/rest/

Thanks!